Heron Page
Added another note to the app and added the page for the Heron note.
Fixed a bug that was reading codes as invalid when the order they are entered didn't exactly match the code array.
Fixed a bug that was showing the wrong number of completed codes on the code list view.